refactor service token
